+++ src/main/java/bridge/controller/RetryController.java
@@ -1,7 +1,6 @@
 package bridge.controller;
 
-import bridge.Lang;
-import bridge.util.Console;
+import java.util.List;
 
 public class RetryController extends Controller {
     public RetryController() {
@@ -13,21 +12,10 @@ public class RetryController extends Controller {
     }
 
     public void printGameMapAndStatus(boolean isPass) {
-        Console.printLine(Lang.get(Lang.GAME_RESULT));
-
         int step = this.getLastTreadStep();
-        this.outputView.printMap(this.getSlapMaps(step));
-        Console.printLine("");
+        int tries = this.getRetries();
+        List<List<String>> maps = this.getSlapMaps(step);
 
-        Console.printLine(Lang.format(Lang.IS_GAME_PASS, this.getGameStatusText(isPass)));
-        Console.printLine(Lang.format(Lang.TOTAL_ATTEMPT, this.getRetries()));
-    }
-
-    private String getGameStatusText(boolean isPass) {
-        if (isPass) {
-            return Lang.get(Lang.GAME_IS_PASS);
-        }
-
-        return Lang.get(Lang.GAME_IS_FAIL);
+        this.outputView.printResult(maps, isPass, tries);
     }
 }
